Refactor RailmlService by extracting RailmlMapper and RailmlTopologyBuilder, Fix DoubleTrack template loading crash
